requirements.txt
asgiref==3.4.1
automium==0.2.6
automium-web==0.1.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
Django==2.0
django-allauth==0.45.0
django-crispy-forms==1.12.0
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
legacy==0.1.6
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python-social-auth==0.2.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
Werkzeug==2.0.1
